Seperate logic from GUI
- FreeCite does not generate a non-used BibtexKeyGenerator anymore
- Introduce ExporterFactoryFactory
- Introduce SavePreferenceFactory
