+++ src/main/java/org/jabref/logic/autosaveandbackup/BackupManager.java
@@ -25,6 +25,7 @@ import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.database.event.BibDatabaseContextChangedEvent;
 import org.jabref.model.database.event.CoarseChangeFilter;
 import org.jabref.preferences.JabRefPreferences;
+import org.jabref.preferences.SavePreferencesFactory;
 
 import com.google.common.eventbus.Subscribe;
 import org.slf4j.Logger;
@@ -119,7 +120,7 @@ public class BackupManager {
     private void performBackup(Path backupPath) {
         try {
             Charset charset = bibDatabaseContext.getMetaData().getEncoding().orElse(preferences.getDefaultEncoding());
-            SavePreferences savePreferences = SavePreferences.loadForSaveFromPreferences(preferences).withEncoding
+            SavePreferences savePreferences = SavePreferencesFactory.loadForSaveFromPreferences(preferences).withEncoding
                     (charset).withMakeBackup(false);
             new BibtexDatabaseWriter<>(FileSaveSession::new).saveDatabase(bibDatabaseContext, savePreferences).commit
                     (backupPath);
